Remove use of traits
