+++ phpstan-baseline.neon
@@ -211,11 +211,6 @@ parameters:
 			path: src/Exception/ConnectException.php
 
 		-
-			message: "#^Method GuzzleHttp\\\\Exception\\\\ConnectException\\:\\:setHandlerContext\\(\\) has parameter \\$handlerContext with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Exception/ConnectException.php
-
-		-
 			message: "#^Property GuzzleHttp\\\\Exception\\\\ConnectException\\:\\:\\$handlerContext type has no value type specified in iterable type array\\.$#"
 			count: 1
 			path: src/Exception/ConnectException.php
@@ -232,11 +227,6 @@ parameters:
 
 		-
 			message: "#^Method GuzzleHttp\\\\Exception\\\\RequestException\\:\\:__construct\\(\\) has parameter \\$handlerContext with no value type specified in iterable type array\\.$#"
-			count: 1
-			path: src/Exception/RequestException.php
-
-		-
-			message: "#^Method GuzzleHttp\\\\Exception\\\\RequestException\\:\\:setHandlerContext\\(\\) has parameter \\$handlerContext with no value type specified in iterable type array\\.$#"
 			count: 1
 			path: src/Exception/RequestException.php
