+++ src/Exception/RequestException.php
@@ -1,8 +1,6 @@
 <?php
 namespace GuzzleHttp\Exception;
 
-use GuzzleHttp\Exception\Traits\HandlerContextAwareTrait;
-use GuzzleHttp\Exception\Traits\RequestAwareTrait;
 use Psr\Http\Client\RequestExceptionInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -13,11 +11,14 @@ use Psr\Http\Message\UriInterface;
  */
 class RequestException extends TransferException implements RequestExceptionInterface
 {
-    use RequestAwareTrait;
-    use HandlerContextAwareTrait;
+    /** @var RequestInterface */
+    private $request;
 
     /** @var ResponseInterface|null */
     private $response;
+
+    /** @var array */
+    private $handlerContext;
 
     public function __construct(
         string $message,
@@ -29,9 +30,9 @@ class RequestException extends TransferException implements RequestExceptionInte
         // Set the code of the exception if the response is set and not future.
         $code = $response ? $response->getStatusCode() : 0;
         parent::__construct($message, $code, $previous);
-        $this->setRequest($request);
+        $this->request = $request;
         $this->response = $response;
-        $this->setHandlerContext($handlerContext);
+        $this->handlerContext = $handlerContext;
     }
 
     /**
@@ -118,6 +119,14 @@ class RequestException extends TransferException implements RequestExceptionInte
     }
 
     /**
+     * Get the request that caused the exception
+     */
+    public function getRequest(): RequestInterface
+    {
+        return $this->request;
+    }
+
+    /**
      * Get the associated response
      */
     public function getResponse(): ?ResponseInterface
@@ -131,5 +140,18 @@ class RequestException extends TransferException implements RequestExceptionInte
     public function hasResponse(): bool
     {
         return $this->response !== null;
+    }
+
+    /**
+     * Get contextual information about the error from the underlying handler.
+     *
+     * The contents of this array will vary depending on which handler you are
+     * using. It may also be just an empty array. Relying on this data will
+     * couple you to a specific handler, but can give more debug information
+     * when needed.
+     */
+    public function getHandlerContext(): array
+    {
+        return $this->handlerContext;
     }
 }
